feat: reorganise routes
